+++ src/sql/__init__.py
@@ -1,7 +1,7 @@
 from sql.magic import load_ipython_extension
 
 
-__version__ = "0.10.11dev"
+__version__ = "0.10.13dev"
 
 
 __all__ = ["load_ipython_extension"]